requirements.txt
selenium>=4.0.0
pytest>=7.0.0
allure-pytest>=2.13.0
webdriver-manager>=3.8.0 